# perf: replace object locks with SemaphoreSlim for improved performance in high concurrent scenarios